Fixed the issue when some apps were gone and couldn't be installed successfully after updating from KK to M via OTA.
There is more than one signature for some apps like Cortana. This app
which is installed in KK would be gone when updating from KK to M via OTA.
The reason for this is the unmatching signature, the error code is:
INSTALL_FAILED_UPDATE_INCOMPATIBLE.

Solution:
'databaseVersion' should not be assigned to CURRENT_DATABASE_VERSION
when creating a 'VersionInfo' instance, because it would affect the
process of ‘compareSignaturesCompat’ signature verification. We could
set the right value of 'databaseVersion' at an appropriate timing.

Signed-off-by: longhai <longhai@xiaomi.com>