Fix slow anim in custom launchers
When using a custom launcher, like Nova, it is possible to set a gesture to open the status bar pulldown. It does not open at the same speed as when using a finder to pull down from the top of the screen at the status bar. 

Credits to SlimROM team

Author Dave Kover <kover@cyngn.com> Jul 1, 2016 7:19 PM
Committer Josue Rivera <prbassplayer@gmail.com> Aug 16, 2016 3:43 PM