Only call uncrypt when we're encrypted
We were always generating a block map for encrypted update, even
when the device was not encrypted.  This leads to a spectacular
failure.  Fix this by only calling uncrypt when we're encrypted.
Additionally, only pass block.map as the update file in the case
that the device was encrypted and requires it.

Pat Erley committed with rogersb11 on May 20

Credits to Cyanidel
https://github.com/CyanideL/android_frameworks_base/commit/44809d68ba1de8e9912e642bcc69e9c1b92f7896