BT: Always update service object in BluetoothDevice Class
Problem: In case service down callback is missed, the service
object is not reset. This causes the object to miss the update
when Bluetooth Service again comes up causing Dead Object
Exception when used.

Cause: When the apps using BluetoothDevice Class are inactive
or killed, they can miss the ServiceDown callback. Because of
NULL check, the object is not updated on receiving the service
up callback.

Fix: Removed the condition to always update the Service object
on receiving the Bluetooth Service Up callback.

CRs-Fixed: 1113447

Change-Id: I8cf0f316fb4f8a1df17b7e5476b0cb3bd580fff1