Fix issue that stack remove issue.
When remove stack, it first call DisplayContent.detachStack which will remove stack in mStacks.
And then it will call stack.detachDisplay(), which will remove all the windows in this stack.
it may call displaycontent.moveStack(), in the removing procedure.
In the moveStack, it first remove the stack in mStack, which is already removed in detachStack.
the moveStack should return here but not, and it add the stack into the mStack again.

This fix add check for the above symptom. if stack already be removed, jsut return

Change-Id: Ife67a50277106908f8b4054f2b49ab7f45127e6a

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>